requirements.txt
# Core dependencies
agno>=2.4.0
cohere>=5.0.0
lancedb>=0.25.0

# Environment management
python-dotenv>=1.0.0

# File monitoring (for dropbox auto-ingestion)
watchdog>=3.0.0

# PDF processing
pypdf>=3.0.0

# Semantic chunking (required for SemanticChunking strategy)
chonkie[semantic]>=0.1.0

# Gmail toolkit dependencies (for Agno GmailTools)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Web server
fastapi>=0.100.0
uvicorn>=0.20.0

# Data validation
pydantic>=2.0.0

# Date/time handling
python-dateutil>=2.8.0

# Tabulate for CLI output
tabulate>=0.9.0

# OpenAI SDK
openai>=1.0.0
